Stores updates List View and company contact

+++ Implementation/MISServices/CompanyService.cs
@@ -443,7 +443,7 @@ namespace MPC.Implementation.MISServices
         {
             return cmsPageRepository.Find(pageId);
         }
-        public CompanyResponse GetCompanyById(int companyId)
+        public CompanyResponse GetCompanyById(long companyId)
         {
             return companyRepository.GetCompanyById(companyId);
         }
